pyproject.toml
[project]
name = "hookshot"
version = "0.1.0"
description = "GitHub webhook receiver that runs shell commands from a YAML config"
requires-python = ">=3.10"
dependencies = ["pyyaml>=6.0"]

[project.scripts]
hookshot = "hookshot.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hookshot"]
